Load model and textures
Now we can load the model and display that in the window. The code still
doesn't quite work because we don't update light related uniforms per
frame, we don't have a depth buffer, we don't read the material
parameters from the mtl file, and we read the texture vertically flipped.

* Add object renderer with phong shaders, object material. The client
  will need to first create an object material, then create an object
  with vertex data and the material.
* Add code to load model stored in wavefront obj file into vertex
  buffer.
* Add code to load texture from png image.
* Modify how multiple uniforms are supported: we now ask the Material to
  create descriptor sets on their own instead of doing that in
  mesh_renderer::Renderer.